webapp_requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
transformers
peft
torch
sentence-transformers
pandas
numpy
scikit-learn
openpyxl
xlrd
accelerate
bitsandbytes
python-docx
psutil
seaborn
matplotlib 
